# swagger.json
{
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "http://localhost:4444",
      "description": "Local Server"
    }
  ],
  "info": {
    "version": "0.0.1",
    "title": "lms-backend",
    "description": "",
    "license": {
      "name": "UNLICENSED"
    }
  },
  "paths": {
...
    "/v1/course/{courseId}/chapter": {
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChapterDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChapterCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterDto"
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          },
          "404": {
            "description": "course not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IErrorResponse404"
                }
              }
            }
          }
        }
      }
    },
    "/v1/course/{courseId}/chapter/{id}": {
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/ChapterDto"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartialChapterDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterDto"
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          },
          "404": {
            "description": "chapter not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IErrorResponse404"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterDto"
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          },
          "404": {
            "description": "chapter not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IErrorResponse404"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TypeGuardErrorany": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "expected": {
            "type": "string"
          },
          "value": {},
          "fake_expected_typed_value_": {},
          "name": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "stack": {
            "type": "string"
          }
        },
        "required": [
          "method",
          "expected",
          "value",
          "name",
          "message"
        ]
      },
      "IErrorResponse404": {
        "type": "object",
        "properties": {
          "statusCode": {
            "const": 404
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message",
          "timestamp",
          "path"
        ]
      },
...
      "ChapterDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "courseId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "sequence": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "courseId",
          "title",
          "description",
          "sequence"
        ]
      },
      "ChapterCreateDto": {
        "type": "object",
        "properties": {
          "description": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "title": {
            "type": "string"
          },
          "sequence": {
            "type": "integer"
          }
        },
        "required": [
          "description",
          "title",
          "sequence"
        ]
      },
      "PartialChapterDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "courseId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "sequence": {
            "type": "integer"
          }
        },
        "description": "Make all properties in T optional"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
...
}